app/rag/requirements.txt
Flask==3.0.0
gunicorn==20.1.0
Werkzeug==3.0.1
gradio>=3.36.1
google-cloud-aiplatform==1.36.0
google-cloud-logging
langchain==0.0.327
unstructured
chromadb
markdown
opencv-python
unstructured[pdf]

